feat: add distribution file
